backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
